+++ src/LieGroups.jl
@@ -62,8 +62,13 @@ include("group_operations/multiplication_operation_abelian.jl")
 
 # Actions
 include("group_actions/group_action_interface.jl")
+include("group_actions/addition_action.jl")
 include("group_actions/group_operation_action.jl")
 include("group_actions/multiplication_action.jl")
+include("group_actions/rotation_around_axis_action.jl")
+# These might use some of the above so we include them second
+include("group_actions/columnwise_action.jl")
+include("group_actions/rowwise_action.jl")
 
 # Meta Lie groups
 include("groups/power_group.jl")
@@ -90,6 +95,7 @@ include("groups/special_orthogonal_group.jl")
 
 # Products of Groups
 include("groups/special_euclidean_group.jl")
+include("groups/special_galilean_group.jl")
 
 export AbstractLieGroup
 export LieGroup, LieAlgebra
@@ -110,15 +116,18 @@ export PowerGroupOperation, ProductGroupOperation
 export LeftSemidirectProductGroupOperation, RightSemidirectProductGroupOperation
 
 #
-#1,
+#
 # Group Actions
 export AbstractGroupActionType
 export AbstractLeftGroupActionType, AbstractRightGroupActionType
+export AdditionGroupAction
 export LeftGroupOperationAction, RightGroupOperationAction
 export GroupAction, GroupOperationAction
 export LeftMultiplicationGroupAction
 export InverseLeftGroupOperationAction, InverseRightGroupOperationAction
 export AbstractActionActsOnType, ActionActsOnLeft, ActionActsOnRight
+export RotationAroundAxisAction
+export ColumnwiseGroupAction, RowwiseGroupAction
 
 #
 #
@@ -128,6 +137,7 @@ export GeneralLinearGroup
 export HeisenbergGroup
 export OrthogonalGroup
 export SpecialEuclideanGroup, SpecialLinearGroup
+export SpecialGalileanGroup
 export SpecialOrthogonalGroup, SpecialUnitaryGroup
 export SymplecticGroup
 export TranslationGroup
@@ -167,6 +177,7 @@ export identity_element, identity_element!, is_identity, inner
 export inv, inv!, diff_inv, diff_inv!
 export injectivity_radius
 export jacobian_conjugate, jacobian_conjugate!
+export jacobian_exp, jacobian_exp!
 export lie_bracket, lie_bracket!, log, log!
 export manifold_dimension
 export norm, number_of_coordinates