+++ src/group_actions/group_action_interface.jl
@@ -215,7 +215,7 @@ This can be performed in-place of `q`.
 function apply end
 @doc "$(_doc_apply)"
 function apply(A::GroupAction, g, p)
-    q = allocate_result(base_manifold(A), apply, p, g)
+    q = allocate_result(base_manifold(A), apply, p)
     apply!(A, q, g, p)
     return q
 end
@@ -266,7 +266,7 @@ Compute the differential ``$(_math(:D))_p α_g(p): T_p$(_math(:M)) → T_{σ_g(p
 function diff_apply end
 @doc "$(_doc_diff_apply)"
 function diff_apply(A::GroupAction, g, p, X)
-    Y = allocate_result(base_manifold(A), diff_apply, p, g, X)
+    Y = allocate_result(base_manifold(A), diff_apply, p, X)
     diff_apply!(A, Y, g, p, X)
     return Y
 end
@@ -287,7 +287,7 @@ and for a right action ``τ_p(g) = τ(g,p)``.
 function diff_group_apply end
 @doc "$(_doc_diff_group_apply)"
 function diff_group_apply(A::GroupAction, g, p, X)
-    Y = allocate_result(base_manifold(A), apply, g, p, X)
+    Y = allocate_result(base_manifold(A), apply, g, X)
     diff_group_apply!(A, Y, g, p, X)
     return Y
 end