+++ test/LieGroupsTestSuite.jl/LieGroupsTestSuite.jl
@@ -5,7 +5,9 @@ This module provides tools and dummy structures to test functionality provided
 within `LieGroups.jl`.
 
 For every test function, several interactions to other functions can be activated.
+
 The following functions are expected to be available, since their defaults just pass through to the manifold
+
 * `is_point` both on the Lie group `G` and the Lie algebra `𝔤`
 * `isapprox(G, g, h)` and `isapprox(𝔤, X, Y)`
 * `copy(G, g)`
@@ -692,7 +694,7 @@ function test_identity_element(
             e = identity_element(G, identity_type)
         end
         if !ismissing(expected_value)
-            @test injectivity_radius(G) == expected_value
+            @test identity_element(G) == expected_value
         end
         if test_mutating
             e2 = copy(G, e)
@@ -939,6 +941,40 @@ function test_jacobian_conjugate(
         return nothing
     end
 end
+
+"""
+    test_jacobian_exp(
+        G::AbstractLieGroup, g, X;
+        basis = DefaultLieAlgebraOrthogonalBasis(),
+        expected = missing,
+        test_mutating = true,
+        kwargs...
+    )
+
+test `jacobian_exp`.
+
+"""
+function test_jacobian_exp(
+        G::AbstractLieGroup, g, X;
+        basis = DefaultLieAlgebraOrthogonalBasis(),
+        expected = missing,
+        test_mutating::Bool = true,
+        kwargs...,
+    )
+    @testset "Jacobian of the exponential map with respect to its argument" begin
+        J = jacobian_exp(G, g, X, basis)
+        n = manifold_dimension(base_manifold(G))
+        @test size(J) == (n, n)
+        if test_mutating
+            J2 = copy(J)
+            jacobian_exp!(G, J2, g, X, basis)
+            @test isapprox(J, J2; kwargs...)
+        end
+        !ismissing(expected) && (@test isapprox(J, expected; kwargs...))
+    end
+    return nothing
+end
+
 #
 #
 # --- L
@@ -1137,7 +1173,7 @@ end
 """
     test_lie_group(G::AbstractLieGroup, properties::Dict, expectations::Dict)
 
-Test the Lie group ``G`` based on a `Dict` of properties and a `Dict` of `expectations
+Test the [`AbstractLieGroup`](@ref) ``G`` based on a `Dict` of properties and a `Dict` of `expectations`.
 
 Possible properties are
 
@@ -1153,10 +1189,10 @@ Possible properties are
 
 Possible `expectations` are
 
-* `:adjoint` for the result of `conjgate` in the case where `diff_conjugate` is not implemented
+* `:adjoint` for the result of `'conjugate` in the case where `diff_conjugate` is not implemented
 * `:atol => 0.0` a global absolute tolerance
 * `:atols -> Dict()` a dictionary `function -> atol` for specific function tested.
-* `:conjugate` for the result of `conjgate
+* `:conjugate` for the result of `conjugate
 * `:conjugate_default => false` to activate the test of the default implementation of `conjugate`
 * `:diff_inv` for the result of `diff_inv` with respect to the first point and the first vector.
 * `:diff_left_compose` for the result of `diff_left_compose` with respect to the first two points and the first vector.
@@ -1357,7 +1393,10 @@ function test_lie_group(G::AbstractLieGroup, properties::Dict, expectations::Dic
                 G, points[1], points[2]; expected = v, test_mutating = mutating
             )
         end
-
+        if (jacobian_exp in functions)
+            expected = get(expectations, :jacobian_exp, missing)
+            test_jacobian_exp(G, points[1], vectors[1]; expected = expected)
+        end
         #
         #
         # --- L
@@ -1410,9 +1449,9 @@ function test_lie_group(G::AbstractLieGroup, properties::Dict, expectations::Dic
 end
 
 """
-    test_group_action(G::AbstractLieGroup, properties::Dict, expectations::Dict)
+    test_group_action(A::GroupAction, properties::Dict, expectations::Dict)
 
-Test the Lie group ``G`` based on a `Dict` of properties and a `Dict` of `expectations`.
+Test the [`GroupAction`](@ref) ``A`` based on a `Dict` of `properties` and a `Dict` of `expectations`.
 
 Possible properties are